#1530: Removed circular import.
